Move all automod commands from main bot script to a cog
